ronny raygun

+++ headers/int.h
@@ -7,7 +7,7 @@ typedef signed short int16;
 typedef unsigned short uint16;
 typedef signed int int32;
 typedef unsigned int uint32;
-typedef int32 bool_t;
+typedef uint8 bool_t;
 
 #ifdef __GNUC__
